src/ingestion_worker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
msal==1.29.0
openai==1.35.7
azure-servicebus==7.12.2
azure-eventhub==5.11.7
azure-storage-blob==12.20.0
azure-core==1.30.2
